Use the AlertLabelOptions from AlertProps and remove tests of antd behavior
