+++ src/controls/AlertControl.tsx
@@ -1,11 +1,9 @@
 import { LabelProps, RendererProps } from "@jsonforms/core"
-import { Alert } from "antd"
+import { Alert, AlertProps } from "antd"
 
-export type AlertLabelOptions = {
-    type?: "warning" | "success" | "info"
-  }
+export type AlertLabelOptions = Pick<AlertProps, "type">
 
 export function AlertControl({ text, uischema }: LabelProps & RendererProps) {
-  const options: AlertLabelOptions = uischema?.options as AlertLabelOptions
+  const options = uischema?.options as AlertLabelOptions
   return <Alert style={{ marginBottom: "24px" }} type={options?.type} message={text} showIcon />
 }
